+++ src/pages/Events.tsx
@@ -1,12 +1,37 @@
-import React from "react";
-import EventsList from "../components/events/events-list";
-import EventsTitle from "../components/events/events-title";
+import React, { useEffect, useState } from "react";
+import { Box, Divider } from "@mui/material";
+import PageTitle from "../components/page-title/page-title";
+import { Actions } from "../actions/actions";
+import { Event } from "../models/Event";
+import EventsSection from "../components/events/events-section";
 
 const Events = (): React.ReactElement => {
+  const [events, setEvents] = useState<Event[]>([]);
+
+  useEffect(() => {
+    Actions.getEvents()
+      .then((events) => setEvents(events ?? [])) // o operador ?? diz que se events for nulo, use []
+      .catch((err) => console.error(err));
+  }, []);
+
+  const openEvents: Event[] = events.filter((e) => e.active);
+  const closedEvents: Event[] = events.filter((e) => !e.active);
+
   return (
     <div className="container">
-      <EventsTitle />
-      <EventsList />
+      <Box
+        sx={{
+          display: "flex",
+          justifyContent: "flex-end",
+        }}
+      >
+        <PageTitle text="Eventos" orientation="flex-start" />
+      </Box>
+      <Box marginTop={3}>
+        <EventsSection title="Eventos abertos" data={openEvents} />
+        <Divider sx={{ marginY: "3rem" }} />
+        <EventsSection title="Eventos fechados" data={closedEvents} />
+      </Box>
     </div>
   );
 };
